source/clients/requirements.txt
python-socketio[client]==5.10.0
websocket-client==1.6.4